Initial commit to address issue #194 Centralize Fixtures
